feat(tauri.js) move exported api types into api modules (fix #807)
