+++ .changes/taurijs_move_api_types_into_api.md
@@ -1,0 +1,7 @@
+---
+"tauri.js": minor
+---
+
+Move types exported in the `tauri` js api into the modules that use them. For
+example, `Event` is now available from `tauri/api/event` instead of
+`tauri/api/types/event`.
