--- cli/tauri.js/api-src/types/cli.ts
@@ -1,22 +1,0 @@
-export interface ArgMatch {
-  /**
-   * string if takes value
-   * boolean if flag
-   * string[] or null if takes multiple values
-   */
-  value: string | boolean | string[] | null
-  /**
-   * number of occurrences
-   */
-  occurrences: number
-}
-
-export interface SubcommandMatch {
-  name: string
-  matches: CliMatches
-}
-
-export interface CliMatches {
-  args: { [name: string]: ArgMatch }
-  subcommand: SubcommandMatch | null
-}
