--- cli/tauri.js/api-src/types/dialog.ts
@@ -1,8 +1,0 @@
-export interface OpenDialogOptions {
-  filter?: string
-  defaultPath?: string
-  multiple?: boolean
-  directory?: boolean
-}
-
-export type SaveDialogOptions = Pick<OpenDialogOptions, 'filter' | 'defaultPath'>
