--- cli/tauri.js/api-src/types/event.ts
@@ -1,6 +1,0 @@
-export interface Event<T> {
-  type: string
-  payload: T
-}
-
-export type EventCallback<T> = (event: Event<T>) => void
