--- cli/tauri.js/api-src/types/notification.ts
@@ -1,8 +1,0 @@
-export interface Options {
-  title: string
-  body?: string
-  icon?: string
-}
-
-export type PartialOptions = Omit<Options, 'title'>
-export type Permission = 'granted' | 'denied' | 'default'
